requirements.txt
gradio
groq
gtts
python-dotenv
Pillow
requests
